Moved web pages and adjusted file paths
