+++ BookApp/frontend/Pages/bind.jsx
@@ -1,6 +1,6 @@
 import { useState } from "react";
 import { useNavigate } from "react-router-dom";
-import logo from "../assets/images/logo.svg";
+import logo from "./assets/images/logo.svg";
 
 function BindPage() {
