+++ BookApp/frontend/Pages/homepage.jsx
@@ -1,9 +1,9 @@
 import { useState, useEffect, useRef } from "react";
 import { useNavigate } from "react-router-dom";
-import logo from "../assets/images/logo.svg";
-import bindimage from "../assets/images/bind-image.svg";
-import Logout from "./logout.jsx";
-import Footer from "./footer.jsx";
+import logo from "./assets/images/logo.svg";
+import bindimage from "./assets/images/bind-image.svg";
+import Logout from "./src/components/logout.jsx";
+import Footer from "./src/components/footer.jsx";
 
 function HomePage() {
   const [searchQuery, setSearchQuery] = useState("");
